GrandeProblema

+++ GraphicalObject.h
@@ -8,7 +8,7 @@ class GraphicalObject
 public:
 	GraphicalObject(Graphics* gfx, ImageFile* imgFile, float OffsetX=0, float OffsetY=0, float ScaleX=1,float ScaleY=1,
 	float RotationAngle=0);
-	void SetUVcord(float LowerBoundX, float HigherBoundX, float LowerBoundY, float HigherBoundY);
+	void SetUVcord(int LowerBoundX, int HigherBoundX, int LowerBoundY, int HigherBoundY);
 	void Move(float OffsetX = 0, float OffsetY = 0);
 	void Scale(float ScaleX = 1, float ScaleY = 1);
 	void Rotate(float RotationAngle);
